requirements.txt
easyocr 
PyMuPDF
pillow 
python-docx 
pandas 
openpyxl 
tqdm
